fix(gitlab): Update ID types for GitLab API v1 compatibility
The GitLab API client v1 changed all ID types from `int` to `int64`.
This updates the GitLab forge integration to match,
affecting merge request IDs, note IDs, user IDs, and repository IDs.

Changes include updating type signatures for:
- Service interface methods (`mergeRequestsService`, `notesService`)
- Repository fields (`repoID`, `userID`)
- Helper functions (`assigneeIDs`, `resolveReviewerIDs`, `mergeAssigneeIDs`)
- Data structures (`MR`, `MRComment`)
- String conversion calls (using `strconv.FormatInt` instead of `strconv.Itoa`)
- Type casts where int values are passed to int64 parameters

+++ internal/forge/gitlab/change_meta.go
@@ -87,7 +87,7 @@ func (*Forge) UnmarshalChangeID(data json.RawMessage) (forge.ChangeID, error) {
 type MR struct {
 	// Number is the merge request number.
 	// This will always be set.
-	Number int `json:"number"` // required
+	Number int64 `json:"number"` // required
 }
 
 var _ forge.ChangeID = (*MR)(nil)
